media-source/MediaSource-addSourceBuffer_readyState.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>readyState attribute not in the "open" state</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>
<script>
var video = document.querySelector('video');
var mimes = ['video/webm; codecs="vorbis,vp8"', 'video/mp4'];

//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function test(mime) {
    async_test(function(t) {
        if(!canPlayType.bind(t)(mime)) {
            return;
        }
        var mediaSource = new MediaSource();
        t.step(function() {
            assert_throws({name : 'InvalidStateError'},
                          function() {
                              var sourceBuffer = mediaSource.addSourceBuffer(mime);
                          },
                          'Set invalid duration');
        });
        t.done();
    }, 'MediaSource#addSourceBuffer (' + mime + ') : ' +
       'If the readyState attribute is not in the "open" state then throw an INVALID_STATE_ERR exception and abort these steps.');
}
mimes.forEach(function(mime) {
    test(mime);
});
</script>
</body>
</html>
